Enable locale fallback if translation is missing
Using the locale function now adds the specified language to the front of a
list of languages which are searched for a translation string.

This allows a primary language as well as one or more fallback languages. On
initialization, the languages array contains only English---making English the
default fallback.

To clear the existing languages array upon specifying a language, set the
clearLangList argument of the locale function to true.